+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Release Candidate ##
+
+Changes from [v0.1.0](https://github.com/hifirworks/hifir4m/releases/tag/v0.1.0):
+
+- Extended the `M` parameter of `gmresHif` to allow a function handle as input.
+
+[Full Changelog](https://github.com/hifirworks/hifir4m/compare/v0.1.0...HEAD)
+
 ## [v0.1.0](https://github.com/hifirworks/hifir4m/releases/tag/v0.1.0) (2021-09-02) ##
 
 This is the initial official release of the HIFIR4M package based on the C++ HIFIR [v0.1.0](https://github.com/hifirworks/hifir/releases/tag/v0.1.0).